Add an extension for persisting Hashes to files
Building on @maxlinc's start, this implements a Persistable extension
that you can mix into any Hash to gain access to file persistence.
Currently, there are two adapters: JSON and YAML.

The persistence framework is easily expandable via adapters that can be
registered and hotloaded into a Persistable module. The module defaults
to a JSON adapter with a `#persist` method when mixed into a derived
Hash class, but can be customized as follows:

```ruby
class YamlPersistedHash < Hash
  include Hashie::Extensions::Persistence.new(
    adapter: :yaml,
    persist_method: :save
  )
end

test = YamlPersistedHash['test' => 'value']  #=> {'test' => 'value'}
test.save('filename.yaml')
```

To create a new adapter, you only need to register a class, module, or
object that responds to an `#adapter` method. The `#adapter` method
needs to respond with an object with a `#write(target, data)` method
that knows how to write the data for persistence. For two examples, see
the `Hashie::Extensions::Persistable::Json` and
`Hashie::Extensions::Persistable::Yaml` modules.

Closes #262 by finishing and expanding the implementation.